requirements.txt
boto3
PyYAML
beautifulsoup4
requests
selenium
webdriver_manager
lxml
prettytable
